Tratamento de erros + health checker (actuator) (estado da API e BD)

+++ src/main/java/com/egitron/gestaopedidos/repository/ErrorLogRepository.java
@@ -3,5 +3,4 @@ package com.egitron.gestaopedidos.repository;
 import com.egitron.gestaopedidos.model.ErrorLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ErrorLogRepository extends JpaRepository<ErrorLog, Long> {
-}
+public interface ErrorLogRepository extends JpaRepository<ErrorLog, Integer> {}
